Associate dotenv language with .env files
Instead of associating a list of file extensions starting with `.env`
with the `dotenv` language, The `dotenv` language is now associated with
the `.env*` filename pattern. This also removes the need to write this
file association to the user configuration.

The `.flaskenv` file is now associated via the `.flaskenv` filename
instead of a file extension. Additionally the `user-dirs.dirs` file is
associated with the `dotenv` language. This is an env file used to
configure the XDG user directories.

File extensions I’m not familiar with, are left untouched.